fix: split player actions * one remote and one for tests
