+++ bot/src/main/java/de/hatoka/poker/bot/remote/client/RemotePlayer.java
@@ -1,6 +1,5 @@
 package de.hatoka.poker.bot.remote.client;
 
-import de.hatoka.poker.table.PlayerActions;
 import de.hatoka.poker.table.PlayerGameInfo;
 
 public interface RemotePlayer extends PlayerActions, PlayerGameInfo
