+++ table/src/test/java/de/hatoka/poker/table/internal/event/PlayerGameInfoTest.java
@@ -15,7 +15,6 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import de.hatoka.poker.base.Card;
-import de.hatoka.poker.table.PlayerActions;
 import de.hatoka.poker.table.capi.business.SeatRef;
 import de.hatoka.poker.table.capi.event.game.DealerInGame;
 import de.hatoka.poker.table.capi.event.game.PlayerFactory;
