+++ table/src/test/java/de/hatoka/poker/table/internal/event/PlayerTest.java
@@ -14,7 +14,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import de.hatoka.poker.table.PlayerActions;
 import de.hatoka.poker.table.capi.business.SeatRef;
 import de.hatoka.poker.table.capi.event.IllegalGameEventException;
 import de.hatoka.poker.table.capi.event.game.DealerInGame;
